consolidate preconditioner flags for benchmarks
