+++ benchmark/solver/solver.cpp
@@ -47,6 +47,7 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include "benchmark/utils/general.hpp"
 #include "benchmark/utils/loggers.hpp"
 #include "benchmark/utils/overhead_linop.hpp"
+#include "benchmark/utils/preconditioners.hpp"
 #include "ginkgo/core/preconditioner/isai.hpp"
 
 
@@ -68,24 +69,6 @@ DEFINE_string(
     solvers, "cg",
     "A comma-separated list of solvers to run. "
     "Supported values are: bicgstab, bicg, cg, cgs, fcg, gmres, overhead");
-
-DEFINE_string(
-    preconditioners, "none",
-    "A comma-separated list of preconditioners to use. "
-    "Supported values are: none, jacobi, adaptive-jacobi, parict, parilu, "
-    "parilut, ilu, parict-isai, parilu-isai, parilut-isai, ilu-isai, overhead");
-
-DEFINE_uint32(parilu_iterations, 5,
-              "The number of iterations for ParICT/ParILU(T)");
-
-DEFINE_bool(parilut_approx_select, true,
-            "Use approximate selection for ParICT/ParILUT");
-
-DEFINE_double(parilut_limit, 2.0, "The fill-in limit for ParICT/ParILUT");
-
-DEFINE_int32(
-    isai_power, 1,
-    "Which power of the sparsity structure to use for ISAI preconditioners");
 
 DEFINE_uint32(
     nrhs, 1,
@@ -185,215 +168,6 @@ const std::map<std::string, std::function<std::unique_ptr<gko::LinOpFactory>(
                             .on(exec);
                     }},
                    {"overhead", create_solver<gko::Overhead<>>}};
-
-
-// TODO: Workaround until GPU matrix conversions are implemented
-//       The factory will wrap another factory, and make sure that the
-//       input operator is copied to the reference executor, and then sent
-//       through the generate function
-struct ReferenceFactoryWrapper
-    : gko::EnablePolymorphicObject<ReferenceFactoryWrapper, gko::LinOpFactory> {
-    ReferenceFactoryWrapper(std::shared_ptr<const gko::Executor> exec)
-        : gko::EnablePolymorphicObject<ReferenceFactoryWrapper,
-                                       gko::LinOpFactory>(exec)
-    {}
-
-    ReferenceFactoryWrapper(std::shared_ptr<const gko::LinOpFactory> f)
-        : gko::EnablePolymorphicObject<ReferenceFactoryWrapper,
-                                       gko::LinOpFactory>(f->get_executor()),
-          base_factory{f}
-    {}
-
-    std::shared_ptr<const gko::Executor> exec{gko::ReferenceExecutor::create()};
-    std::shared_ptr<const gko::LinOpFactory> base_factory;
-
-protected:
-    std::unique_ptr<gko::LinOp> generate_impl(
-        std::shared_ptr<const gko::LinOp> op) const override
-    {
-        return base_factory->generate(gko::clone(exec, op));
-    }
-};
-
-
-const std::map<std::string, std::function<std::unique_ptr<gko::LinOpFactory>(
-                                std::shared_ptr<const gko::Executor>)>>
-    precond_factory{
-        {"none",
-         [](std::shared_ptr<const gko::Executor> exec) {
-             return gko::matrix::IdentityFactory<>::create(exec);
-         }},
-        {"jacobi",
-         [](std::shared_ptr<const gko::Executor> exec) {
-             std::shared_ptr<const gko::LinOpFactory> f =
-                 gko::preconditioner::Jacobi<>::build().on(exec);
-             return std::unique_ptr<ReferenceFactoryWrapper>(
-                 new ReferenceFactoryWrapper(f));
-         }},
-        {"adaptive-jacobi",
-         [](std::shared_ptr<const gko::Executor> exec) {
-             std::shared_ptr<const gko::LinOpFactory> f =
-                 gko::preconditioner::Jacobi<>::build()
-                     .with_storage_optimization(
-                         gko::precision_reduction::autodetect())
-                     .on(exec);
-             return std::unique_ptr<ReferenceFactoryWrapper>(
-                 new ReferenceFactoryWrapper(f));
-         }},
-        {"parict",
-         [](std::shared_ptr<const gko::Executor> exec) {
-             auto fact = std::shared_ptr<gko::LinOpFactory>(
-                 gko::factorization::ParIct<>::build()
-                     .with_iterations(FLAGS_parilu_iterations)
-                     .with_approximate_select(FLAGS_parilut_approx_select)
-                     .with_fill_in_limit(FLAGS_parilut_limit)
-                     .on(exec));
-             std::shared_ptr<const gko::LinOpFactory> f =
-                 gko::preconditioner::Ilu<>::build()
-                     .with_factorization_factory(fact)
-                     .on(exec);
-             return std::unique_ptr<ReferenceFactoryWrapper>(
-                 new ReferenceFactoryWrapper(f));
-         }},
-        {"parilu",
-         [](std::shared_ptr<const gko::Executor> exec) {
-             auto fact = std::shared_ptr<gko::LinOpFactory>(
-                 gko::factorization::ParIlu<>::build()
-                     .with_iterations(FLAGS_parilu_iterations)
-                     .on(exec));
-             std::shared_ptr<const gko::LinOpFactory> f =
-                 gko::preconditioner::Ilu<>::build()
-                     .with_factorization_factory(fact)
-                     .on(exec);
-             return std::unique_ptr<ReferenceFactoryWrapper>(
-                 new ReferenceFactoryWrapper(f));
-         }},
-        {"parilut",
-         [](std::shared_ptr<const gko::Executor> exec) {
-             auto fact = std::shared_ptr<gko::LinOpFactory>(
-                 gko::factorization::ParIlut<>::build()
-                     .with_iterations(FLAGS_parilu_iterations)
-                     .with_approximate_select(FLAGS_parilut_approx_select)
-                     .with_fill_in_limit(FLAGS_parilut_limit)
-                     .on(exec));
-             std::shared_ptr<const gko::LinOpFactory> f =
-                 gko::preconditioner::Ilu<>::build()
-                     .with_factorization_factory(fact)
-                     .on(exec);
-             return std::unique_ptr<ReferenceFactoryWrapper>(
-                 new ReferenceFactoryWrapper(f));
-         }},
-        {"ilu",
-         [](std::shared_ptr<const gko::Executor> exec) {
-             auto fact = std::shared_ptr<gko::LinOpFactory>(
-                 gko::factorization::Ilu<>::build().on(exec));
-             std::shared_ptr<const gko::LinOpFactory> f =
-                 gko::preconditioner::Ilu<>::build()
-                     .with_factorization_factory(fact)
-                     .on(exec);
-             return std::unique_ptr<ReferenceFactoryWrapper>(
-                 new ReferenceFactoryWrapper(f));
-         }},
-        {"parict-isai",
-         [](std::shared_ptr<const gko::Executor> exec) {
-             auto fact = std::shared_ptr<gko::LinOpFactory>(
-                 gko::factorization::ParIct<>::build()
-                     .with_iterations(FLAGS_parilu_iterations)
-                     .with_approximate_select(FLAGS_parilut_approx_select)
-                     .with_fill_in_limit(FLAGS_parilut_limit)
-                     .on(exec));
-             auto lisai = gko::share(gko::preconditioner::LowerIsai<>::build()
-                                         .with_sparsity_power(FLAGS_isai_power)
-                                         .on(exec));
-             auto uisai = gko::share(gko::preconditioner::UpperIsai<>::build()
-                                         .with_sparsity_power(FLAGS_isai_power)
-                                         .on(exec));
-             std::shared_ptr<const gko::LinOpFactory> f =
-                 gko::preconditioner::Ilu<
-                     gko::preconditioner::LowerIsai<>,
-                     gko::preconditioner::UpperIsai<>>::build()
-                     .with_factorization_factory(fact)
-                     .with_l_solver_factory(lisai)
-                     .with_u_solver_factory(uisai)
-                     .on(exec);
-             return std::unique_ptr<ReferenceFactoryWrapper>(
-                 new ReferenceFactoryWrapper(f));
-         }},
-        {"parilu-isai",
-         [](std::shared_ptr<const gko::Executor> exec) {
-             auto fact = std::shared_ptr<gko::LinOpFactory>(
-                 gko::factorization::ParIlu<>::build()
-                     .with_iterations(FLAGS_parilu_iterations)
-                     .on(exec));
-             auto lisai = gko::share(gko::preconditioner::LowerIsai<>::build()
-                                         .with_sparsity_power(FLAGS_isai_power)
-                                         .on(exec));
-             auto uisai = gko::share(gko::preconditioner::UpperIsai<>::build()
-                                         .with_sparsity_power(FLAGS_isai_power)
-                                         .on(exec));
-             std::shared_ptr<const gko::LinOpFactory> f =
-                 gko::preconditioner::Ilu<
-                     gko::preconditioner::LowerIsai<>,
-                     gko::preconditioner::UpperIsai<>>::build()
-                     .with_factorization_factory(fact)
-                     .with_l_solver_factory(lisai)
-                     .with_u_solver_factory(uisai)
-                     .on(exec);
-             return std::unique_ptr<ReferenceFactoryWrapper>(
-                 new ReferenceFactoryWrapper(f));
-         }},
-        {"parilut-isai",
-         [](std::shared_ptr<const gko::Executor> exec) {
-             auto fact = std::shared_ptr<gko::LinOpFactory>(
-                 gko::factorization::ParIlut<>::build()
-                     .with_iterations(FLAGS_parilu_iterations)
-                     .with_approximate_select(FLAGS_parilut_approx_select)
-                     .with_fill_in_limit(FLAGS_parilut_limit)
-                     .on(exec));
-             auto lisai = gko::share(gko::preconditioner::LowerIsai<>::build()
-                                         .with_sparsity_power(FLAGS_isai_power)
-                                         .on(exec));
-             auto uisai = gko::share(gko::preconditioner::UpperIsai<>::build()
-                                         .with_sparsity_power(FLAGS_isai_power)
-                                         .on(exec));
-             std::shared_ptr<const gko::LinOpFactory> f =
-                 gko::preconditioner::Ilu<
-                     gko::preconditioner::LowerIsai<>,
-                     gko::preconditioner::UpperIsai<>>::build()
-                     .with_factorization_factory(fact)
-                     .with_l_solver_factory(lisai)
-                     .with_u_solver_factory(uisai)
-                     .on(exec);
-             return std::unique_ptr<ReferenceFactoryWrapper>(
-                 new ReferenceFactoryWrapper(f));
-         }},
-        {"ilu-isai",
-         [](std::shared_ptr<const gko::Executor> exec) {
-             auto fact = std::shared_ptr<gko::LinOpFactory>(
-                 gko::factorization::Ilu<>::build().on(exec));
-             auto lisai = gko::share(gko::preconditioner::LowerIsai<>::build()
-                                         .with_sparsity_power(FLAGS_isai_power)
-                                         .on(exec));
-             auto uisai = gko::share(gko::preconditioner::UpperIsai<>::build()
-                                         .with_sparsity_power(FLAGS_isai_power)
-                                         .on(exec));
-             std::shared_ptr<const gko::LinOpFactory> f =
-                 gko::preconditioner::Ilu<
-                     gko::preconditioner::LowerIsai<>,
-                     gko::preconditioner::UpperIsai<>>::build()
-                     .with_factorization_factory(fact)
-                     .with_l_solver_factory(lisai)
-                     .with_u_solver_factory(uisai)
-                     .on(exec);
-             return std::unique_ptr<ReferenceFactoryWrapper>(
-                 new ReferenceFactoryWrapper(f));
-         }},
-        {"overhead", [](std::shared_ptr<const gko::Executor> exec) {
-             std::shared_ptr<const gko::LinOpFactory> f =
-                 gko::Overhead<>::build().on(exec);
-             return std::unique_ptr<ReferenceFactoryWrapper>(
-                 new ReferenceFactoryWrapper(f));
-         }}};
 
 
 void write_precond_info(const gko::LinOp *precond,
